# openapi.yaml
openapi: 3.0.3
info:
  title: NeXFlowX Core Banking API
  version: 5.3.1-dev
  description: API Institucional (Fonte Única de Verdade) para gestão de Wallets, Swaps, Payouts e Liquidez.
servers:
  - url: https://api-dev.nexflowx.tech/api/v1
    description: Sandbox Environment
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Autenticação de Utilizador
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: "XDeals"
                password:
                  type: string
                  example: "xdeals123"
      responses:
        '200':
          description: Login com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      username:
                        type: string
                      role:
                        type: string
                        enum: [admin, merchant, customer]
...
  /payment-links:
    post:
      tags: [Payments]
      summary: Gerar link de depósito (Top-up)
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, currency]
              properties:
                amount:
                  oneOf:
                    - type: number
                    - type: string
                currency:
                  type: string
      responses:
        '201':
          description: Link de checkout gerado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      shareable_url:
                        type: string